Move consent remove action to consent controller
The action moved from the deprovision controller to the consent
controller. And the routing_api.yml file was updated accordingly.

In doing so, to prevent creating duplicate code. The factory logic for creating
the CollabPersonId was extracted into a little factory helper class.
