+++ src/OpenConext/EngineBlockBundle/Controller/Api/ConsentController.php
@@ -21,6 +21,8 @@ namespace OpenConext\EngineBlockBundle\Controller\Api;
 use OpenConext\EngineBlock\Exception\RuntimeException;
 use OpenConext\EngineBlock\Service\ConsentServiceInterface;
 use OpenConext\EngineBlockBundle\Configuration\FeatureConfigurationInterface;
+use OpenConext\EngineBlockBundle\Exception\InvalidArgumentException as EngineBlockInvalidArgumentException;
+use OpenConext\EngineBlockBundle\Factory\CollabPersonIdFactory;
 use OpenConext\EngineBlockBundle\Http\Exception\ApiAccessDeniedHttpException;
 use OpenConext\EngineBlockBundle\Http\Exception\ApiInternalServerErrorHttpException;
 use OpenConext\EngineBlockBundle\Http\Exception\ApiMethodNotAllowedHttpException;
@@ -88,5 +90,44 @@ final class ConsentController
         }
 
         return new JsonResponse($consentList, Response::HTTP_OK);
+    }
+
+    public function removeAction(Request $request): JsonResponse
+    {
+        if (!$request->isMethod(Request::METHOD_POST)) {
+            throw ApiMethodNotAllowedHttpException::methodNotAllowed($request->getMethod(), [Request::METHOD_POST]);
+        }
+
+        if (!$this->featureConfiguration->isEnabled('api.consent_remove')) {
+            throw new ApiNotFoundHttpException('Consent remove API is disabled');
+        }
+
+        if (!$this->authorizationChecker->isGranted('ROLE_API_USER_PROFILE')) {
+            throw new ApiAccessDeniedHttpException(
+                'Access to the consent removal API requires the role ROLE_API_USER_PROFILE'
+            );
+        }
+
+        $userId = $request->get('collabPersonId', false);
+        $serviceProviderEntityId = $request->get('serviceProviderEntityId', false);
+        if (!$userId || !$serviceProviderEntityId) {
+            throw new EngineBlockInvalidArgumentException('The required data for removing the consent is not present in the request parameters');
+        }
+
+        try {
+            $user = CollabPersonIdFactory::create($userId);
+            $removed = $this->deprovisionService->deleteOneConsentFor($user, $serviceProviderEntityId);
+        } catch (RuntimeException $e) {
+            throw new ApiInternalServerErrorHttpException(
+                sprintf(
+                    'An unknown error occurred while removing a service the user has given consent for to ' .
+                    'release attributes to ("%s")',
+                    $e->getMessage()
+                ),
+                $e
+            );
+        }
+
+        return new JsonResponse($removed, Response::HTTP_OK);
     }
 }
