+++ src/OpenConext/EngineBlockBundle/Factory/CollabPersonIdFactory.php
@@ -1,0 +1,43 @@
+<?php
+/**
+ * Copyright 2010 SURFnet B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace OpenConext\EngineBlockBundle\Factory;
+
+use InvalidArgumentException;
+use OpenConext\EngineBlock\Authentication\Value\CollabPersonId;
+use OpenConext\EngineBlockBundle\Http\Exception\ApiNotFoundHttpException;
+use function sprintf;
+
+class CollabPersonIdFactory
+{
+    /**
+     * @throws ApiNotFoundHttpException
+     */
+    public static function create(string $id): CollabPersonId
+    {
+        try {
+            return new CollabPersonId($id);
+        } catch (InvalidArgumentException $e) {
+            throw new ApiNotFoundHttpException(
+                sprintf(
+                    'User ID is not valid: "%s"',
+                    $e->getMessage()
+                )
+            );
+        }
+    }
+}
